:tada: Suspense is now limited to within search, allowing button to render on mount
