+++ aksel.nav.no/website/app/_ui/global-search/GlobalSearch.provider.tsx
@@ -2,58 +2,25 @@
 
 import {
   createContext,
-  useCallback,
   useContext,
   useEffect,
   useMemo,
   useRef,
   useState,
-  useTransition,
 } from "react";
-import { debounce } from "@navikt/ds-react";
-import { useParamState } from "@/app/_ui/global-search/useParamState";
-import { umamiTrack } from "@/app/_ui/umami/Umami.track";
-import { fuseGlobalSearch } from "./GlobalSearch.actions";
-
-type ActionReturnT = Awaited<ReturnType<typeof fuseGlobalSearch>>;
 
 type SearchContextType = {
   open: boolean;
   openSearch: () => void;
   closeSearch: () => void;
   inputRef: React.MutableRefObject<HTMLInputElement | null>;
-  queryResults: ActionReturnT | null;
-  updateQuery: (query: string) => void;
-  resetSearch: () => void;
 };
 
 const SearchContext = createContext<SearchContextType | null>(null);
 
 function GlobalSearchProvider({ children }: { children: React.ReactNode }) {
   const inputRef = useRef<HTMLInputElement | null>(null);
-  const shouldInitialOpenRef = useRef<boolean>(true);
-
   const [open, setOpen] = useState<boolean>(false);
-  const [searchResult, setSearchResults] = useState<ActionReturnT | null>(null);
-
-  const [, startTransition] = useTransition();
-  const { setParam, clearParam, paramValue } = useParamState("query");
-
-  const debouncedSearch = useMemo(
-    () =>
-      debounce((query: string) => {
-        maybeEnableComicSans(query);
-
-        umamiTrack("sok", {});
-        setParam(query);
-      }, 200),
-    [setParam],
-  );
-
-  const resetSearch = useCallback(() => {
-    debouncedSearch.clear();
-    clearParam();
-  }, [clearParam, debouncedSearch]);
 
   useEffect(() => {
     const listener = (event: KeyboardEvent) => {
@@ -76,40 +43,14 @@ function GlobalSearchProvider({ children }: { children: React.ReactNode }) {
     return () => document.removeEventListener("keydown", listener);
   }, [open]);
 
-  useEffect(() => {
-    if (!paramValue) {
-      setSearchResults(null);
-      shouldInitialOpenRef.current = false;
-
-      return;
-    }
-
-    startTransition(async () => {
-      const newResults = await fuseGlobalSearch(paramValue);
-
-      setSearchResults(newResults);
-
-      /*
-       * We use a ref to ensure the dialog only auto-opens after results are in to avoid CLS.
-       */
-      if (shouldInitialOpenRef.current) {
-        setOpen(true);
-        shouldInitialOpenRef.current = false;
-      }
-    });
-  }, [paramValue]);
-
   const contextValue = useMemo(
     () => ({
       open,
       closeSearch: () => setOpen(false),
       openSearch: () => setOpen(true),
       inputRef,
-      queryResults: searchResult,
-      updateQuery: debouncedSearch,
-      resetSearch,
     }),
-    [open, searchResult, debouncedSearch, resetSearch],
+    [open],
   );
 
   return (
@@ -117,12 +58,6 @@ function GlobalSearchProvider({ children }: { children: React.ReactNode }) {
       {children}
     </SearchContext.Provider>
   );
-}
-
-function maybeEnableComicSans(query: string) {
-  if (query.includes("comic")) {
-    document.body.style.fontFamily = "Comic Sans MS, cursive, sans-serif";
-  }
 }
 
 function useGlobalSearch() {
