+++ aksel.nav.no/website/app/_ui/global-search/GlobalSearch.results.tsx
@@ -2,14 +2,14 @@
 
 import { useEffect } from "react";
 import { BodyShort, Heading } from "@navikt/ds-react";
+import { useGlobalSearchResults } from "@/app/_ui/global-search/GlobalSearch.search-provider";
 import { preloadSearchIndex } from "./GlobalSearch.actions";
 import { globalSearchConfig } from "./GlobalSearch.config";
 import { GlobalSearchHitCollection } from "./GlobalSearch.hit";
 import styles from "./GlobalSearch.module.css";
-import { useGlobalSearch } from "./GlobalSearch.provider";
 
 function GlobalSearchResultsView() {
-  const { queryResults } = useGlobalSearch();
+  const { queryResults } = useGlobalSearchResults();
 
   if (!queryResults?.result || queryResults?.result?.totalHits === 0) {
     return null;
@@ -53,7 +53,7 @@ function GlobalSearchPreload() {
 }
 
 function GlobalSearchEmptySearchState() {
-  const { queryResults } = useGlobalSearch();
+  const { queryResults } = useGlobalSearchResults();
 
   const showEmptySearchState =
     !queryResults?.result?.totalHits && queryResults?.query;
@@ -157,7 +157,7 @@ function GlobalSearchEmptySearchState() {
 }
 
 function GlobalSearchEmptyState() {
-  const { queryResults } = useGlobalSearch();
+  const { queryResults } = useGlobalSearchResults();
 
   if (queryResults?.result) {
     return null;
