+++ aksel.nav.no/website/app/_ui/header/Header.tsx
@@ -1,6 +1,5 @@
 import clsx from "clsx";
 import Link from "next/link";
-import { Suspense } from "react";
 import { Box, HStack, Hide, Show, Spacer } from "@navikt/ds-react";
 import { GlobalSearch } from "@/app/_ui/global-search/GlobalSearch";
 import { MobileNav } from "@/app/_ui/mobile-nav/MobileNav";
@@ -71,9 +70,8 @@ function Header({ variant }: { variant?: "default" | "produktbloggen" }) {
           </Box>
         </Show>
         <HStack align="center" gap="2">
-          <Suspense>
-            <GlobalSearch />
-          </Suspense>
+          <GlobalSearch />
+
           <Show below="lg">
             <MobileNav />
           </Show>
